fix: remove dev preview and plotly refs
