+++ README.md
@@ -1,6 +1,6 @@
 # 🎰 Casino Event Calendar
 
-A personal Dash application that displays casino events on a responsive calendar. Weekly and daily views include interactive modals and an optional CSS grid preview.
+A personal Dash application that displays casino events on a responsive calendar. Weekly and daily views include interactive modals rendered with a CSS grid layout.
 
 ---
 
@@ -8,8 +8,7 @@ A personal Dash application that displays casino events on a responsive calendar
 
 - Weekly calendar view with color‑coded event blocks
 - Modal windows for detailed event and day information
-- Built-in CSS grid layout preview
-- Toggle to show or hide the Plotly weekly grid
+- CSS grid layout for weekly view
 - Responsive design for desktop, tablet and mobile
 - Time zone normalized to Pacific Time (PDT)
 - Toggle to show ongoing events that span the week
@@ -27,7 +26,7 @@ app_components/          # Core logic modules
   layout.py              # Layout and modals
   plotting.py            # Plotly figure generation
   utils.py               # Helpers and time zone utilities
-  week_grid_layout.py    # Grid layout preview using HTML/CSS
+  week_grid_layout.py    # CSS-based week grid layout
 assets/                  # Static assets auto-loaded by Dash
   base.css
   styles/
