+++ TODO.md
@@ -1,0 +1,5 @@
+## CSS Grid Parity TODO
+
+- Add hover tooltips on grid events similar to Plotly hover info.
+- Allow clicking on empty day areas to open the day view.
+- Refine mobile scaling and text truncation for very small screens.
